module forcage + fix
