+++ py4cast/datasets/dummy.py
@@ -23,6 +23,8 @@ from py4cast.datasets.base import (
     TorchDataloaderSettings,
     collate_fn,
 )
+
+
 from py4cast.plots import DomainInfo
 from py4cast.settings import CACHE_DIR
