+++ py4cast/datasets/poesy.py
@@ -24,6 +24,9 @@ from py4cast.datasets.base import (
     TorchDataloaderSettings,
     collate_fn,
 )
+
+from py4cast.forcingutils.utils import get_year_hour_forcing, generate_toa_radiation_forcing 
+
 from py4cast.plots import DomainInfo
 from py4cast.settings import CACHE_DIR
 
@@ -470,13 +473,13 @@ class PoesyDataset(DatasetABC, Dataset):
         sample = self.sample_list[index]
 
         # Datetime Forcing
-        datetime_forcing = self.get_year_hour_forcing(
+        datetime_forcing = get_year_hour_forcing(
             sample.date, sample.output_terms
         ).type(torch.float32)
 
         # Solar forcing, dim : [num_pred_steps, Lat, Lon, feature = 1]
-        solar_forcing = self.generate_toa_radiation_forcing(
-            self.grid, sample.date, sample.output_terms
+        solar_forcing = generate_toa_radiation_forcing(
+            self.grid.lat, self.grid.lon, sample.date, sample.output_terms
         ).type(torch.float32)
 
         lforcings = [
